refactor userMgr, getUser, update file structure
